dashboard sidebar update
