+++ apps/dashboard/src/features/dashboard/components/header.tsx
@@ -1,0 +1,61 @@
+import {
+  Breadcrumb,
+  BreadcrumbItem,
+  BreadcrumbLink,
+  BreadcrumbList,
+  BreadcrumbPage,
+  BreadcrumbSeparator,
+} from "@workspace/ui/components/breadcrumb";
+import { Separator } from "@workspace/ui/components/separator";
+import { SidebarTrigger } from "@workspace/ui/components/sidebar";
+
+import { NotificationsPopover } from "./nav-notifications";
+
+const sampleNotifications = [
+  {
+    id: "1",
+    avatar: "/avatars/01.png",
+    fallback: "OM",
+    text: "New order received.",
+    time: "10m ago",
+  },
+  {
+    id: "2",
+    avatar: "/avatars/02.png",
+    fallback: "JL",
+    text: "Server upgrade completed.",
+    time: "1h ago",
+  },
+  {
+    id: "3",
+    avatar: "/avatars/03.png",
+    fallback: "HH",
+    text: "New user signed up.",
+    time: "2h ago",
+  },
+];
+
+export function Header() {
+  return (
+    <header className="flex h-16 shrink-0 items-center justify-between gap-2 px-4 transition-[width,height] ease-linear group-has-[[data-collapsible=icon]]/sidebar-wrapper:h-12">
+      <div className="flex items-center gap-2">
+        <SidebarTrigger className="-ml-1" />
+        <Separator orientation="vertical" className="mr-2 h-4" />
+        <Breadcrumb>
+          <BreadcrumbList>
+            <BreadcrumbItem className="hidden md:block">
+              <BreadcrumbLink href="#">
+                Building Your Application
+              </BreadcrumbLink>
+            </BreadcrumbItem>
+            <BreadcrumbSeparator className="hidden md:block" />
+            <BreadcrumbItem>
+              <BreadcrumbPage>Data Fetching</BreadcrumbPage>
+            </BreadcrumbItem>
+          </BreadcrumbList>
+        </Breadcrumb>
+      </div>
+      <NotificationsPopover notifications={sampleNotifications} />
+    </header>
+  );
+}
