+++ apps/dashboard/src/features/dashboard/index.ts
@@ -1,1 +1,2 @@
 export { AppSidebar } from "./components/app-sidebar";
+export { Header } from "./components/header";
